Implement JSON parsing functionality and related structures
- Added `jsonshard.rs` for JSON parsing logic, including object and value matching.
- Introduced `leaves.rs` to define grammar matching for string and character types.
- Created `mod.rs` to organize shard modules and expose necessary components.
- Developed `numbers.rs` to handle numeric parsing, including integers, unsigned integers, hexadecimals, and reals.
- Established `parser.rs` to define the `Parser` struct and the `IGrammar` trait for grammar matching.
- Implemented `primeshard.rs` for whitespace handling in parsing.
- Added `repeatshard.rs` to support repetition in parsing constructs.
- Created `shardtree.rs` for constructing parsing trees using macros.
- Developed `node.rs` to define binary and unary nodes for expression trees.

+++ src/flux/jsonoutstrm.rs
@@ -2,7 +2,7 @@
 use std::{fmt, mem::swap};
 
 use crate::flux::fluxexport::{FieldExp, IFluxExportSink};
-use u32;
+use std::u32;
 
 //---------------------------------------------------------------------------------------------------------------------------------
 
@@ -19,7 +19,7 @@ impl<W: fmt::Write> JsonOutStream<W> {
     pub fn New(ostr: W, multiLineFlg: bool) -> Self {
         Self {
             _OStr: ostr,
-            _Depth: 0,
+            _Depth: (0 as u32),
             _EntryFlg: false,
             _MultiLineFlg: multiLineFlg,
         }
